# docs/specs/data-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Outage Explorer Data API",
    "version": "1.0.0",
    "description": "Implemented HTTP contract for explicitly configured services. Synthetic examples; Linux runtime validation belongs to the user and live resources are not enabled."
  },
  "paths": {
    "/api/datasets": {
      "get": {
        "operationId": "listDatasets",
        "summary": "listDatasets",
        "x-implementation-status": "implemented",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Catalog",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalog"
                }
              }
            }
          },
          "401": {
            "description": "Error401",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error401"
                }
              }
            }
          },
          "403": {
            "description": "Error403",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error403"
                }
              }
            }
          },
          "503": {
            "description": "Error503",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                },
                "description": "Bounded retry suggestion, agrees with error.retry_after_seconds when present."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error503"
                }
              }
            }
          },
          "400": {
            "description": "Error400",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error400"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "description": "Requires current session and application authorization. All responses use no-store. Duplicate/unknown parameters and unsupported bodies are rejected."
      }
    },
...
    "/api/query": {
      "post": {
        "operationId": "executeQuery",
        "summary": "executeQuery",
        "x-implementation-status": "implemented",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "QueryResult",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResult"
                }
              }
            }
          },
          "400": {
            "description": "Error400",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error400"
                }
              }
            }
          },
          "401": {
            "description": "Error401",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error401"
                }
              }
            }
          },
          "403": {
            "description": "Error403",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error403"
                }
              }
            }
          },
          "422": {
            "description": "Error422",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error422"
                }
              }
            }
          },
          "429": {
            "description": "Error429",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                },
                "description": "Bounded retry suggestion, agrees with error.retry_after_seconds when present."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error429"
                }
              }
            }
          },
          "503": {
            "description": "Error503",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                },
                "description": "Bounded retry suggestion, agrees with error.retry_after_seconds when present."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error503"
                }
              }
            }
          },
          "504": {
            "description": "Error504",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error504"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "description": "Requires current session and application authorization. All responses use no-store. Duplicate/unknown parameters and unsupported bodies are rejected. SQL is executed only by POST and is never modified for pagination. GET reads an owned immutable result. Default page size 100, maximum 500; expiry is 15 minutes from execution completion.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getQueryPage",
        "summary": "getQueryPage",
        "x-implementation-status": "implemented",
        "parameters": [
          {
            "name": "query_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500
            }
          }
        ],
        "responses": {
          "200": {
            "description": "QueryResult",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResult"
                }
              }
            }
          },
          "400": {
            "description": "Error400",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error400"
                }
              }
            }
          },
          "401": {
            "description": "Error401",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error401"
                }
              }
            }
          },
          "403": {
            "description": "Error403",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error403"
                }
              }
            }
          },
          "404": {
            "description": "Error404",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error404"
                }
              }
            }
          },
          "410": {
            "description": "Error410",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error410"
                }
              }
            }
          },
          "503": {
            "description": "Error503",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                },
                "description": "Bounded retry suggestion, agrees with error.retry_after_seconds when present."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error503"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "description": "Requires current session and application authorization. All responses use no-store. Duplicate/unknown parameters and unsupported bodies are rejected. SQL is executed only by POST and is never modified for pagination. GET reads an owned immutable result. Default page size 100, maximum 500; expiry is 15 minutes from execution completion."
      }
    },
    "/api/refresh": {
      "post": {
        "operationId": "admitRefresh",
        "summary": "admitRefresh",
        "x-implementation-status": "implemented",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{16,128}$"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "RefreshReceipt",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshReceipt"
                }
              }
            }
          },
          "400": {
            "description": "Error400",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error400"
                }
              }
            }
          },
          "401": {
            "description": "Error401",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error401"
                }
              }
            }
          },
          "403": {
            "description": "Error403",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error403"
                }
              }
            }
          },
          "409": {
            "description": "Error409",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                },
                "description": "Bounded retry suggestion, agrees with error.retry_after_seconds when present."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error409"
                }
              }
            }
          },
          "503": {
            "description": "Error503",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                },
                "description": "Bounded retry suggestion, agrees with error.retry_after_seconds when present."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error503"
                }
              }
            }
          },
          "202": {
            "description": "RefreshReceipt",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              },
              "Location": {
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "schema": {
                  "const": "3"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshReceipt"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "description": "Requires current session and application authorization. All responses use no-store. Duplicate/unknown parameters and unsupported bodies are rejected.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-outage_session",
        "description": "Production HTTPS cookie. Explicit local HTTP mode uses outage_session; reuse the configured auth transport."
      }
    },
    "schemas": {
...
      "Catalog": {
        "type": "object",
        "properties": {
          "generation_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "datasets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Dataset"
            }
          }
        },
        "required": [
          "generation_id",
          "datasets"
        ],
        "additionalProperties": false
      },
...
      "QueryResult": {
        "type": "object",
        "properties": {
          "query_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "generation_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ]
          },
          "columns": {
            "$ref": "#/components/schemas/Columns"
          },
          "rows": {
            "$ref": "#/components/schemas/Rows"
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "page_size": {
            "type": "integer",
            "minimum": 1,
            "maximum": 500
          },
          "retained_row_count": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000
          },
          "total_pages": {
            "type": "integer",
            "minimum": 1
          },
          "has_more": {
            "type": "boolean"
          },
          "truncated": {
            "type": "boolean"
          },
          "truncation_reason": {
            "enum": [
              null,
              "row_limit",
              "byte_limit"
            ]
          },
          "limits": {
            "$ref": "#/components/schemas/Limits"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "Z$"
          }
        },
        "required": [
          "query_id",
          "generation_id",
          "columns",
          "rows",
          "page",
          "page_size",
          "retained_row_count",
          "total_pages",
          "has_more",
          "truncated",
          "truncation_reason",
          "limits",
          "expires_at"
        ],
        "additionalProperties": false
      },
      "QueryRequest": {
        "type": "object",
        "properties": {
          "sql": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "sql"
        ],
        "additionalProperties": false
      },
...
      "Error400": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "enum": [
                  "invalid_request",
                  "invalid_sql",
                  "unsupported_sql",
                  "page_out_of_range",
                  "page_size_mismatch"
                ]
              },
              "message": {
                "type": "string"
              },
              "retry_after_seconds": {
                "type": "integer",
                "minimum": 1,
                "maximum": 3600
              },
              "details": {
                "type": "object",
                "properties": {
                  "query_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "Z$"
                  }
                },
                "required": [
                  "query_id",
                  "expires_at"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "Error401": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "enum": [
                  "unauthenticated"
                ]
              },
              "message": {
                "type": "string"
              },
              "retry_after_seconds": {
                "type": "integer",
                "minimum": 1,
                "maximum": 3600
              },
              "details": {
                "type": "object",
                "properties": {
                  "query_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "Z$"
                  }
                },
                "required": [
                  "query_id",
                  "expires_at"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "Error403": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "enum": [
                  "forbidden"
                ]
              },
              "message": {
                "type": "string"
              },
              "retry_after_seconds": {
                "type": "integer",
                "minimum": 1,
                "maximum": 3600
              },
              "details": {
                "type": "object",
                "properties": {
                  "query_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "Z$"
                  }
                },
                "required": [
                  "query_id",
                  "expires_at"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "Error422": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "enum": [
                  "query_resource_limit"
                ]
              },
              "message": {
                "type": "string"
              },
              "retry_after_seconds": {
                "type": "integer",
                "minimum": 1,
                "maximum": 3600
              },
              "details": {
                "type": "object",
                "properties": {
                  "query_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "Z$"
                  }
                },
                "required": [
                  "query_id",
                  "expires_at"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "Error429": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "enum": [
                  "result_capacity_exhausted"
                ]
              },
              "message": {
                "type": "string"
              },
              "retry_after_seconds": {
                "type": "integer",
                "minimum": 1,
                "maximum": 3600
              },
              "details": {
                "type": "object",
                "properties": {
                  "query_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "Z$"
                  }
                },
                "required": [
                  "query_id",
                  "expires_at"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "Error503": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "enum": [
                  "query_busy",
                  "result_capacity_exhausted",
                  "data_unavailable",
                  "service_unavailable"
                ]
              },
              "message": {
                "type": "string"
              },
              "retry_after_seconds": {
                "type": "integer",
                "minimum": 1,
                "maximum": 3600
              },
              "details": {
                "type": "object",
                "properties": {
                  "query_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "Z$"
                  }
                },
                "required": [
                  "query_id",
                  "expires_at"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
    }
  }
}